Reformat controllers.

+++ controller/admin/delete_file.php
@@ -6,19 +6,19 @@ $db->setAttribute(PDO::ATTR_ERRMODE, PDO::ERRMODE_EXCEPTION);
 function delete_file($file_id) {
     global $db;
     
-    try {
+    try 
+    {
         $db->beginTransaction();
         
         $stmt = $db->prepare("SELECT path FROM Files WHERE id = ?");
         $stmt->execute([$file_id]);
         $file = $stmt->fetch(PDO::FETCH_ASSOC);
         
-        if ($file) {
+        if ($file) 
+        {
             $file_path = $file['path'];
             
-            if (file_exists($file_path)) {
-                unlink($file_path);
-            }
+            if (file_exists($file_path)) { unlink($file_path); }
 
             $stmt = $db->prepare("DELETE FROM Sectors WHERE file_id = ?");
             $stmt->execute([$file_id]);
@@ -33,18 +33,17 @@ function delete_file($file_id) {
             
             echo "File and associated records deleted successfully.";
             header("Location: ../../vue/admin/files_list.php");
-        } else {
-            echo "File not found.";
-        }
-    } catch (Exception $e) {
+        } else { echo "File not found."; }
+    } 
+    catch (Exception $e) 
+    {
         $db->rollBack();
         die('Error: ' . $e->getMessage());
     }
 }
 
-if ($_SERVER["REQUEST_METHOD"] === "POST" && isset($_POST['file_id'])) {
+if ($_SERVER["REQUEST_METHOD"] === "POST" && isset($_POST['file_id'])) 
+{
     $file_id = $_POST['file_id'];
     delete_file($file_id);
-} else {
-    echo "Invalid request.";
-}
+} else { echo "Invalid request."; }
